+++ ananoid.tests/FunctionsTests.fs
@@ -5,9 +5,12 @@
 *)
 module pblasucci.Ananoid.Tests.CoreTests
 
+open global.Xunit
+
 open System
-open FsCheck
-open FsCheck.Xunit
+
+open Hedgehog.FSharp
+open Hedgehog.Xunit
 
 (* ⮟ system under test ⮟ *)
 open pblasucci.Ananoid
@@ -16,34 +19,49 @@ open pblasucci.Ananoid.Core
 open pblasucci.Ananoid.Core.Tagged
 
 
-[<Property>]
-let ``Returns empty on negative size`` (NegativeInt size) =
-  let value = nanoIdOf Alphabets.Numbers size
-  value |> String.IsNullOrWhiteSpace |> Prop.label $"generated '%s{value}'"
-
-[<Property>]
-let ``Raises exception on zero-length alphabet`` (NonNegativeInt count) =
-  lazy (nanoIdOf (String.replicate count " ") 21)
-  |> Prop.throws<ArgumentOutOfRangeException, _>
-
-[<Property(MaxTest = 1)>]
-let ``Raises exception on over-large alphabet`` () =
-  lazy (nanoIdOf (String.replicate 1024 "-") 21)
-  |> Prop.throws<ArgumentOutOfRangeException, _>
-
-[<Property(MaxTest = 1)>]
-let ``Default is UrlSafe alphabet of size 21`` () =
-  let value = nanoId ()
-  let parsed = UrlSafe |> Alphabet.parseNanoId value
-  value.Length = 21 && Option.isSome parsed
-
-[<Property(Arbitrary = [| typeof<Generation> |])>]
-let ``Tagged output equals untagged output`` (TaggedNanoId tagged) =
+[<Property(typeof<Generation>)>]
+let ``Tagged output equals untagged output`` (tagged : string<nanoid>) =
   let untagged = string tagged
   tagged = nanoid.tag untagged
 
-[<Property(MaxTest = 1)>]
+[<Property>]
+let ``Returns empty on negative size`` ([<SmallNegativeInt>] size) =
+  match nanoIdOf Alphabets.Numbers size with
+  | "" -> Property.success ()
+  | value -> Property.counterexample (fun () -> $"generated '%s{value}'")
+
+[<Fact>]
+let ``Raises exception on zero-length alphabet`` () =
+  Assert.Throws<ArgumentOutOfRangeException>(fun () ->
+    let nanoId = nanoIdOf (String.replicate 52 " ") Defaults.Size
+    printfn $"Expected exn, but got: %A{nanoId}"
+  )
+
+[<Fact>]
+let ``Raises exception on over-large alphabet`` () =
+  Assert.Throws<ArgumentOutOfRangeException>(
+    "alphabet",
+    (fun () -> nanoIdOf (String.replicate 1024 "-") 21 |> printfn "%s")
+  )
+
+[<Fact>]
+let ``Default is UrlSafe alphabet of size 21`` () =
+  let value = nanoId ()
+  Assert.Multiple(
+    (fun () -> Assert.Equal(21, value.Length)),
+    (fun () ->
+      // NOTE "can be parsed" is a good-enough proxy for "is of alphabet"
+      Assert.True(UrlSafe |> Alphabet.parseNanoId value |> Option.isSome)
+    )
+  )
+
+[<Fact>]
 let ``Tagged default is UrlSafe alphabet of size 21`` () =
   let value = string (nanoId' ())
-  let parsed = value |> NanoId.parseAs UrlSafe
-  value.Length = 21 && Option.isSome parsed
+  Assert.Multiple(
+    (fun () -> Assert.Equal(21, value.Length)),
+    (fun () ->
+      // NOTE "can be parsed" is a good-enough proxy for "is of alphabet"
+      Assert.True(value |> NanoId.parseAs UrlSafe |> Option.isSome)
+    )
+  )